+++ srunner/scenarioconfigs/route_scenario_configuration.py
@@ -68,7 +68,7 @@ class RouteScenarioConfiguration(ScenarioConfiguration):
         self.ego_vehicles = []
         self.trigger_points = []
 
-        self.name = "RouteScenario"
+        self.name = "RouteScenario_{}".format(route_description['id'])
         self.town = route_description['town_name']
         self.route_description = route_description